+++ docker/Dockerfile.rocm
@@ -178,8 +178,8 @@ RUN git clone https://github.com/radixark/miles.git /root/miles && \
     cd /root/miles && \
     git checkout ${MILES_COMMIT} && \
     git apply --3way --unidiff-zero /tmp/amd_patch/miles.patch && \
-    mkdir -p miles/backends/experimental/fsdp_utils/sglang_attn_bridge && \
-    cp -r /tmp/amd_patch/sglang_attn_bridge/. miles/backends/experimental/fsdp_utils/sglang_attn_bridge/ && \
+    mkdir -p miles/backends/fsdp_utils/sglang_attn_bridge && \
+    cp -r /tmp/amd_patch/sglang_attn_bridge/. miles/backends/fsdp_utils/sglang_attn_bridge/ && \
     pip install -e . --no-deps
 
 # ====================================== Install sgl-model-gateway ============================================
