+++ miles/backends/fsdp_utils/adaptations/specs/glm4_moe_lite.py
@@ -2,7 +2,7 @@
 expert layout matches qwen3_moe, so it reuses the same HF-native unfuse) and the
 rollout-routing-replay hook on its group-limited router."""
 
-from miles.backends.experimental.fsdp_utils.models.replay_routers import install_glm4_moe_lite_router_replay
+from miles.backends.fsdp_utils.models.replay_routers import install_glm4_moe_lite_router_replay
 from ..routing_replay import RoutingReplayAdapter, register_routing_replay_adapter
 from ..weight_bridge import _batched_experts_matches, _hf_unfuse_experts_expand, register_param_transform
