+++ miles/backends/fsdp_utils/parallel.py
@@ -4,11 +4,9 @@ from argparse import Namespace
 import torch.distributed as dist
 from torch.distributed.device_mesh import DeviceMesh, init_device_mesh
 
+from miles.backends.training_utils.parallel import ParallelState
 from miles.utils.distributed_utils import get_gloo_group
-
 from miles.utils.ft_utils.process_group_utils import GroupInfo
-
-from ...training_utils.parallel import ParallelState
 
 logger = logging.getLogger(__name__)
