+++ miles/ray/train/actor_factory.py
@@ -66,7 +66,7 @@ def allocate_gpus_for_actor(
         actor_impl = MegatronTrainRayActor
 
     else:
-        from miles.backends.experimental.fsdp_utils import FSDPTrainRayActor
+        from miles.backends.fsdp_utils import FSDPTrainRayActor
 
         actor_impl = FSDPTrainRayActor
