+++ miles/utils/arguments.py
@@ -2608,7 +2608,7 @@ def parse_args(add_custom_arguments=None):
         args.world_size = args.actor_num_nodes * args.actor_num_gpus_per_node
         args = set_default_megatron_args(args)
     else:
-        from miles.backends.experimental.fsdp_utils.arguments import load_fsdp_args
+        from miles.backends.fsdp_utils.arguments import load_fsdp_args
 
         args = load_fsdp_args(extra_args_provider=add_miles_arguments)
         # TODO: unify this .rank and .world_size w/ indep_dp logics
@@ -2644,7 +2644,7 @@ def parse_args(add_custom_arguments=None):
                 "pipeline_model_parallel_size is 1."
             )
     else:
-        from miles.backends.experimental.fsdp_utils.arguments import validate_hybrid_shard_args
+        from miles.backends.fsdp_utils.arguments import validate_hybrid_shard_args
 
         validate_hybrid_shard_args(args)
